requirements.txt
fastapi
uvicorn[standard]
asyncpg
sqlalchemy
alembic
psycopg2-binary
python-dotenv
telethon
python-multipart 
PyMuPDF 
openai